Moving and breaking
I got pushing from left and right, there is a bug in pushing from top
and bottom related to the borders around the sprites and I believe it is
also preventing pulling from working. The borders around the temp
sprites i made are too big, or the collision system allows objects to
collide with the sides closer than the top and bottom.